Remove the physical device bits and use init_once in the pool getter
