+++ libcudacxx/include/cuda/__memory_pool/device_memory_pool.h
@@ -32,7 +32,11 @@
 #  include <cuda/std/__concepts/concept_macros.h>
 #  include <cuda/std/__memory/construct_at.h>
 #  include <cuda/std/__memory/unique_ptr.h>
-#  include <cuda/std/optional>
+#  include <cuda/std/__type_traits/is_trivially_destructible.h>
+
+#  if _CCCL_HOSTED()
+#    include <mutex>
+#  endif // _CCCL_HOSTED()
 
 #  include <cuda/std/__cccl/prologue.h>
 
@@ -89,29 +93,61 @@ public:
   using default_queries = ::cuda::mr::properties_list<::cuda::mr::device_accessible>;
 };
 
+struct __default_device_memory_pool
+{
+#  if _CCCL_HOSTED()
+  ::std::once_flag __once_{};
+#  else // ^^^ _CCCL_HOSTED() ^^^ / vvv _CCCL_FREESTANDING() vvv
+  bool __initialized_{false};
+#  endif // _CCCL_FREESTANDING()
+
+  union __storage_t
+  {
+    char __empty_;
+    device_memory_pool_ref __pool_;
+
+    _CCCL_HOST_API __storage_t() noexcept
+        : __empty_{}
+    {}
+  } __storage_;
+
+  _CCCL_HOST_API void __init(::cuda::device_ref __device)
+  {
+    ::cuda::std::__construct_at(
+      &__storage_.__pool_,
+      ::cuda::__get_default_memory_pool(
+        ::CUmemLocation{::CU_MEM_LOCATION_TYPE_DEVICE, __device.get()}, ::CU_MEM_ALLOCATION_TYPE_PINNED));
+#  if !_CCCL_HOSTED()
+    __initialized_ = true;
+#  endif // !_CCCL_HOSTED()
+  }
+
+  [[nodiscard]] _CCCL_HOST_API device_memory_pool_ref& __get(::cuda::device_ref __device)
+  {
+#  if _CCCL_HOSTED()
+    ::std::call_once(__once_, [this, __device]() {
+      this->__init(__device);
+    });
+#  else // ^^^ _CCCL_HOSTED() ^^^ / vvv _CCCL_FREESTANDING() vvv
+    if (!__initialized_)
+    {
+      this->__init(__device);
+    }
+#  endif // _CCCL_FREESTANDING()
+    return __storage_.__pool_;
+  }
+};
+
+static_assert(::cuda::std::is_trivially_destructible_v<device_memory_pool_ref>);
+
 //! @brief  Returns the default ``cudaMemPool_t`` from the specified device.
 //! @throws cuda_error if retrieving the default ``cudaMemPool_t`` fails.
 //! @returns The default memory pool of the specified device.
 [[nodiscard]] inline device_memory_pool_ref& device_default_memory_pool(::cuda::device_ref __device)
 {
-  static ::cuda::std::unique_ptr<::cuda::std::optional<device_memory_pool_ref>[]> __pools_ = []() {
-    const size_t __device_count = ::cuda::__physical_devices().size();
-    ::cuda::std::unique_ptr<::cuda::std::optional<device_memory_pool_ref>[]> __pools{
-      static_cast<::cuda::std::optional<device_memory_pool_ref>*>(
-        ::operator new[](sizeof(::cuda::std::optional<device_memory_pool_ref>) * __device_count))};
-    for (size_t __device = 0; __device < __device_count; ++__device)
-    {
-      ::cuda::std::__construct_at(__pools.get() + __device, ::cuda::std::nullopt);
-    }
-    return __pools;
-  }();
-
-  auto& __pool = __pools_[__device.get()];
-  if (!__pool.has_value())
-  {
-    __pool.emplace(::cuda::__physical_devices()[__device.get()].__get_default_memory_pool());
-  }
-  return *__pool;
+  static ::cuda::std::unique_ptr<__default_device_memory_pool[]> __pools_{
+    ::new __default_device_memory_pool[::cuda::__physical_devices_count()]};
+  return __pools_[__device.get()].__get(__device);
 }
 
 //! @rst
